define constant types to help optimizer

+++ src/physics/constants.jl
@@ -1,7 +1,7 @@
-const hbar = 1.05457266e-27 # erg s
-const c = 2.99792458e10 # cm/s
-const eV = 1.60218e-12 # erg
-const MeV = 1e6 * eV
-const GeV = 1e9 * eV
-const G_F = 1.1663787e-5 / GeV^2 * (hbar*c)^3 # erg cm^3
-const kB = 1.3807e-16 # erg/K
+const hbar::Float64 = 1.05457266e-27 # erg s
+const c::Float64 = 2.99792458e10 # cm/s
+const eV::Float64 = 1.60218e-12 # erg
+const MeV::Float64 = 1e6 * eV
+const GeV::Float64 = 1e9 * eV
+const G_F::Float64 = 1.1663787e-5 / GeV^2 * (hbar*c)^3 # erg cm^3
+const kB::Float64 = 1.3807e-16 # erg/K
